# src/main/java/org/traccar/api/resource/EventResource.java
/*
 * Copyright 2016 - 2021 Anton Tananaev (anton@traccar.org)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.traccar.api.resource;

import java.sql.SQLException;
import java.time.Instant;
import java.time.LocalDate;
import java.time.ZoneId;
import java.time.ZoneOffset;
import java.time.temporal.TemporalUnit;
import java.util.*;
import java.util.function.Predicate;
import java.util.regex.Matcher;
import java.util.regex.Pattern;
import java.util.stream.Collectors;

import javax.ws.rs.*;
import javax.ws.rs.core.MediaType;
import javax.ws.rs.core.Response;

import liquibase.pro.packaged.Q;
import org.traccar.Context;
import org.traccar.api.BaseResource;
import org.traccar.model.*;
import org.traccar.model.Calendar;

@Path("events")
@Produces(MediaType.APPLICATION_JSON)
@Consumes(MediaType.APPLICATION_JSON)

public class EventResource extends BaseResource {

    @Path("{id}")
    @GET
    public Event get(@PathParam("id") long id) throws SQLException {
        Event event = Context.getDataManager().getObject(Event.class, id);
        if (event == null) {
            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND).build());
        }
        Context.getPermissionsManager().checkDevice(getUserId(), event.getDeviceId());
        if (event.getGeofenceId() != 0) {
            Context.getPermissionsManager().checkPermission(Geofence.class, getUserId(), event.getGeofenceId());
        }
        if (event.getMaintenanceId() != 0) {
            Context.getPermissionsManager().checkPermission(Maintenance.class, getUserId(), event.getMaintenanceId());
        }
        return event;
    }


    @Path("/overspeed")
    @GET
    public Collection<ExtendedEvent> get(
            @QueryParam("groupIds") List<Long> groupIds,
            @QueryParam("deviceIds") List<Long> deviceIds,
            @QueryParam("from") Date from,
            @QueryParam("to") Date to,
            @QueryParam("geofenceIds") List<Long> geofencesIds,
            @QueryParam("includeOutsideGeofences") boolean includeOutsideGeofences,
            @QueryParam("minGeofenceSpeedLimit") int minGeofenceSpeedLimit,
            @QueryParam("naxGeofenceSpeedLimit") int naxGeofenceSpeedLimit,
            @QueryParam("minDeviceSpeed") int minDeviceSpeed,
            @QueryParam("maxDeviceSpeed") int maxDeviceSpeed,
            @QueryParam("minDeviceSpeedLimit") int minDeviceSpeedLimit,
            @QueryParam("maxDeviceSpeedLimit") int maxDeviceSpeedLimit
            ) throws SQLException {

        if(from == null){
            //System.out.println("No from date provided");
            LocalDate local = LocalDate.now().minusYears(100);
            from = java.sql.Date.from(local.atStartOfDay().toInstant(ZoneOffset.UTC));
        }
        if(to == null){
            //System.out.println("No to date provided");
            LocalDate local = LocalDate.now().plusYears(1);
            to = java.sql.Date.from(local.atStartOfDay().toInstant(ZoneOffset.UTC));
        }

        // Optimize performance of permissions

        Collection<ExtendedEvent>events = Context.getDataManager().getOverspeedEvents(groupIds, deviceIds, from, to, geofencesIds, includeOutsideGeofences, minDeviceSpeed, maxDeviceSpeed,minDeviceSpeedLimit, maxDeviceSpeedLimit,  minGeofenceSpeedLimit, naxGeofenceSpeedLimit);
        events = events.stream().filter(this::hasPermission).collect(Collectors.toList());
        return events;
    }

    private boolean hasPermission(Event event){
        try{
            Context.getPermissionsManager().checkDevice(getUserId(), event.getDeviceId());
            if(event.getGeofenceId() != 0)
                Context.getGeofenceManager().checkItemPermission(getUserId(), event.getGeofenceId());
            return true;
        }catch (Exception e){
            return false;
        }
    }


}